Move Navigation component to its own folder.
